pyproject.toml
[tool.poetry]
name = "auth-service-by-phone-number"
version = "0.1.0"
description = ""
authors = ["Aristov Sergei <maxmit83@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
phonenumbers = "^8.13.40"
django-cron = "^0.6.0"
drf-yasg = "^1.21.7"
setuptools = "^72.1.0"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
